+++ src/applications/financial-status-report/components/FinancialOverview.jsx
@@ -1,101 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-
-const getMonthlyIncome = ({
-  questions,
-  personalData,
-  additionalIncome,
-  socialSecurity,
-  benefits,
-}) => {
-  const { employmentHistory } = personalData;
-  let totalArr = [];
-
-  if (questions.vetIsEmployed) {
-    const { grossMonthlyIncome } = employmentHistory.veteran.currentEmployment;
-    totalArr = [...totalArr, grossMonthlyIncome];
-  }
-
-  if (questions.spouseIsEmployed) {
-    const { grossMonthlyIncome } = employmentHistory.spouse.currentEmployment;
-    totalArr = [...totalArr, grossMonthlyIncome];
-  }
-
-  if (questions.hasAdditionalIncome) {
-    const vetAddl = additionalIncome.additionalIncomeRecords.map(
-      record => record.monthlyIncome,
-    );
-    totalArr = [...totalArr, ...vetAddl];
-  }
-
-  if (questions.spouseHasAdditionalIncome) {
-    const spouseAddl = additionalIncome.spouse.additionalIncomeRecords.map(
-      record => record.monthlyIncome,
-    );
-    totalArr = [...totalArr, ...spouseAddl];
-  }
-
-  if (questions.hasSocialSecurity) {
-    const { socialSecurityAmount } = socialSecurity;
-    totalArr = [...totalArr, socialSecurityAmount];
-  }
-
-  if (questions.spouseHasSocialSecurity) {
-    const { socialSecurityAmount } = socialSecurity.spouse;
-    totalArr = [...totalArr, socialSecurityAmount];
-  }
-
-  if (questions.spouseHasBenefits) {
-    const { benefitAmount, educationAmount } = benefits.spouseBenefits;
-    totalArr = [...totalArr, benefitAmount, educationAmount];
-  }
-
-  return totalArr.reduce((acc, income) => acc + income, 0);
-};
-
-const getMonthlyExpenses = ({
-  questions,
-  personalData,
-  expenses,
-  otherExpenses,
-  utilityRecords,
-  installmentContractsAndOtherDebts,
-}) => {
-  const { employmentHistory } = personalData;
-  let totalArr = [];
-
-  const householdExpenses = Object.values(expenses);
-  totalArr = [...totalArr, ...householdExpenses];
-
-  if (questions.vetIsEmployed) {
-    const { deductions } = employmentHistory.veteran.currentEmployment;
-    const payrollDeductions = deductions.map(
-      deduction => deduction.deductionAmount,
-    );
-    totalArr = [...totalArr, ...payrollDeductions];
-  }
-
-  if (questions.hasUtilities) {
-    const utilities = utilityRecords.map(
-      utility => utility.monthlyUtilityAmount,
-    );
-    totalArr = [...totalArr, ...utilities];
-  }
-
-  if (questions.hasRepayments) {
-    const installments = installmentContractsAndOtherDebts.map(
-      installment => installment.amountDueMonthly,
-    );
-    totalArr = [...totalArr, ...installments];
-  }
-
-  if (questions.hasOtherExpenses) {
-    const other = otherExpenses.map(expense => expense.expenseAmount);
-    totalArr = [...totalArr, ...other];
-  }
-
-  return totalArr.reduce((acc, expense) => acc + expense, 0);
-};
+import { getMonthlyIncome, getMonthlyExpenses } from '../utils/helpers';
 
 const FinancialOverview = ({ formData }) => {
   const { selectedDebts } = formData;
